backend/requirements.txt
flask
waitress
flask_cors
psycopg2-binary
pyjwt
flask_apscheduler